requirements.txt
beautifulsoup
simplejson
python-dateutil
django
docutils
pil
celery
redis
django-redis-cache
south
fabric
haystack
django-tagging
lettuce
nose
django-form-utils
sorl-thumbnail
markdown
django-ckeditor
django-chronograph